Create location service
