+++ client/src/app/location.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { LocationService } from './location.service';
+
+describe('LocationService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [LocationService]
+    });
+  });
+
+  it('should be created', inject([LocationService], (service: LocationService) => {
+    expect(service).toBeTruthy();
+  }));
+});
