+++ client/src/app/main-menu/main-menu.component.html
@@ -1,3 +1,3 @@
 <ul>
-   <li><button mat-raised-button (click)="fetchClosest()">Closest stops</button></li>
+   <li><button mat-raised-button (click)="locationService.fetchClosest()">Closest stops</button></li>
 </ul>
