+++ client/src/app/main-menu/main-menu.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { HttpClient, HttpParams } from '@angular/common/http';
+import { LocationService } from '../location.service';
 
 @Component({
    selector: 'app-main-menu',
@@ -8,30 +8,8 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 })
 export class MainMenuComponent implements OnInit {
 
-   constructor(private http: HttpClient) { }
+   constructor(private locationService: LocationService) { }
 
    ngOnInit() {
-   }
-
-   public fetchClosest() {
-      if (navigator.geolocation) {
-         navigator.geolocation.getCurrentPosition((position) => {
-            console.log(position);
-            this.http.get('stations', {
-               params: new HttpParams({
-                  fromObject: {
-                     longtitude: position.coords.longitude.toString(),
-                     latitude: position.coords.latitude.toString(),
-                     accuracy: position.coords.accuracy.toString(),
-                  }
-               })
-            }).subscribe((result: any): void => {
-               console.log(result);
-            });
-
-         });
-      } else {
-         console.log('Geolocation is not supported by this browser.');
-      }
    }
 }
